chore: Clean up controller.event package in dhis-api [DHIS2-17403]
