+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/trackedentity/TrackedEntityAnalyticsController.java
@@ -66,13 +66,13 @@ import org.hisp.dhis.analytics.trackedentity.TrackedEntityAnalyticsQueryService;
 import org.hisp.dhis.analytics.trackedentity.TrackedEntityQueryParams;
 import org.hisp.dhis.analytics.trackedentity.TrackedEntityQueryRequestMapper;
 import org.hisp.dhis.analytics.trackedentity.TrackedEntityRequestParams;
-import org.hisp.dhis.common.DimensionsCriteria;
 import org.hisp.dhis.common.Grid;
 import org.hisp.dhis.common.OpenApi;
 import org.hisp.dhis.datavalue.DataValue;
 import org.hisp.dhis.security.RequiresAuthority;
 import org.hisp.dhis.webapi.dimension.DimensionFilteringAndPagingService;
 import org.hisp.dhis.webapi.dimension.DimensionMapperService;
+import org.hisp.dhis.webapi.dimension.DimensionsCriteria;
 import org.hisp.dhis.webapi.dimension.TrackedEntityAnalyticsPrefixStrategy;
 import org.hisp.dhis.webapi.mvc.annotation.ApiVersion;
 import org.hisp.dhis.webapi.utils.ContextUtils;
