+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/dimension/DimensionFilteringAndPagingService.java
@@ -45,7 +45,6 @@ import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import lombok.RequiredArgsConstructor;
 import org.hisp.dhis.analytics.dimensions.AnalyticsDimensionsPagingWrapper;
-import org.hisp.dhis.common.DimensionsCriteria;
 import org.hisp.dhis.common.Pager;
 import org.hisp.dhis.fieldfiltering.FieldFilterParams;
 import org.hisp.dhis.fieldfiltering.FieldFilterService;
